+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/Hobrus/hobrusmetrics.git
 
 go 1.22.7
+
+require github.com/stretchr/testify v1.9.0 // indirect
